+++ core/src/test/scala/sbn/core/models/BayesianNetworkSpec.scala
@@ -1,4 +1,4 @@
-package ferjorosa.sbn.core.models
+package sbn.core.models
 
 /**
   * Created by fer on 10/11/16.
