+++ core/src/test/scala/sbn/core/models/graph/DirectedGraphSpec.scala
@@ -1,11 +1,11 @@
-package ferjorosa.sbn.core.models.graph
+package sbn.core.models.graph
 
-import ferjorosa.sbn.core.CustomSpec
-import ferjorosa.sbn.core.io.DataFileLoader
-import ferjorosa.sbn.core.variables.{Variable, VariableFactory}
+import sbn.core.CustomSpec
+import sbn.core.io.DataFileLoader
+import sbn.core.variables.{Variable, VariableFactory}
 
-import scalax.collection.immutable.Graph
 import scalax.collection.GraphEdge.DiEdge
+import scalax.collection.immutable.Graph
 
 class DirectedGraphSpec extends CustomSpec{
